% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmean.r
\name{fillna}
\alias{fillna}
\title{Fill with NA missing data}
\usage{
fillna(x, ...)
}
\arguments{
\item{x}{A xts data}
}
\value{
Data fill NA
}
\description{
\code{fillna} Missing data fill by NA
}
\author{
Mahmut MUSLUM, \email{mmuslumm@gmail.com}
}

