% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmean.r
\name{MeanBy}
\alias{MeanBy}
\title{Calculate Mean by  time units}
\usage{
MeanBy(xdata, on, percNA = NULL)
}
\arguments{
\item{x}{A xts data}
}
\value{
Mean data
}
\description{
\code{MeanBy} Calculate Mean by  time units
}
\author{
Mahmut MUSLUM, \email{mmuslumm@gmail.com}
}

