% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmean.r
\name{Mode}
\alias{Mode}
\title{Finds the mode of xts object.}
\usage{
Mode(x)
}
\arguments{
\item{x}{A xts data}
}
\value{

}
\description{
\code{Mode} Mode of xts object
}
\author{
Mahmut MUSLUM, \email{mmuslumm@gmail.com}
}

